[ThinLTO] Prepare `RemoteModuleCacheEntry` for async
Prior to this patch, the implementation of `RemoteModuleCacheEntry` assumed `tryLoadingBuffer()` always finishes before `write()` gets called. This assumption was leveraged to perform determinism checks in the `write()` function. In the future, these two functions will run concurrently, and the assumption will no longer hold. This patch extracts the determinism check into new function `areLoadedAndWrittenResultsIdentical()` that will be guaranteed to be always called from context where both of those functions have already returned.

This patch also takes care to avoid concurrent writes into single file from those two functions. This is done by manipulating these three files:
* The ".downloaded.tmp" file, where we're downloading the cached buffer.
* The ".computed.tmp" file, where we're streaming out the computed buffer.
* The file at `OutputPath`, where we atomically move either of the above. This file also acts as the backing buffer for the "UseBufferAPI" mode.

Note that there's no synchronization between the atomic moves, so it's possible for `tryLoadingBuffer()` and `getMappedBuffer()` to return the buffer written out by `write()`. In practise, we expect these buffers to be identical (and we verify that with determinism checks) so this should be fine.